+++ src/Tests/VmScaleSet.fs
@@ -325,4 +325,41 @@ let tests =
                 "my-scale-set"
                 "VMSS OS profile has incorrect computer name prefix"
         }
+
+        test "VMSS does not set deleteOption even when delete_attached is used" {
+            let deployment = arm {
+                add_resources [
+                    vmss {
+                        name "my-scale-set"
+                        capacity 2
+
+                        vm_profile (
+                            vm {
+                                username "azureuser"
+                                operating_system UbuntuServer_2204LTS
+                                vm_size Standard_B1s
+                                os_disk 128 StandardSSD_LRS
+                                add_ssd_disk 256
+                                delete_attached // This should be ignored for scale sets
+                            }
+                        )
+                    }
+                ]
+            }
+
+            let jobj = deployment.Template |> Writer.toJson |> JToken.Parse
+            let vmss = jobj.SelectToken("resources[?(@.name=='my-scale-set')]")
+            let vmProfile = vmss.SelectToken("properties.virtualMachineProfile")
+
+            // Check OS disk does NOT have deleteOption
+            let osDiskDeleteOption = vmProfile.SelectToken("storageProfile.osDisk.deleteOption")
+
+            Expect.isNull osDiskDeleteOption "VMSS OS disk should NOT have deleteOption property"
+
+            // Check data disk does NOT have deleteOption
+            let dataDiskDeleteOption =
+                vmProfile.SelectToken("storageProfile.dataDisks[0].deleteOption")
+
+            Expect.isNull dataDiskDeleteOption "VMSS data disk should NOT have deleteOption property"
+        }
     ]